% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttexplore-package.R
\docType{package}
\name{ttexplore-package}
\alias{ttexplore}
\alias{ttexplore-package}
\title{ttexplore: Explore TidyTuesday Datasets}
\description{
The goal of ttexplore is to...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dslc-io/ttexplore}
  \item \url{https://dslc-io.github.io/ttexplore/}
  \item Report bugs at \url{https://github.com/dslc-io/ttexplore/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
